fix genre-list
