+++ src/components/app/app.jsx
@@ -8,15 +8,11 @@ class App extends PureComponent {
   constructor(props) {
     super(props);
 
-    this.state = {
-      selectedFilm: null,
-    };
-
     this.showMoreButtonClick = this.showMoreButtonClick.bind(this);
   }
 
-  showMoreButtonClick() {
-
+  showMoreButtonClick(evt) {
+    evt.preventDefault();
   }
 
   render() {
