+++ src/reducer.js
@@ -1,4 +1,5 @@
-import {extend, Genres} from "./utils.js";
+import {extend} from "./utils.js";
+import {Genres} from "./mocks/consts.js";
 import films from "./mocks/films.js";
 
 const initialState = {
