+++ src/utils.js
@@ -2,17 +2,4 @@ const extend = (a, b) => {
   return Object.assign({}, a, b);
 };
 
-const Genres = {
-  ALL: `All genres`,
-  COMEDIES: `Comedies`,
-  CRIME: `Crime`,
-  DOCUMENTARY: `Documentary`,
-  DRAMA: `Drama`,
-  HORROR: `Horror`,
-  KIDS_FAMILY: `Kids & Family`,
-  ROMANCE: `Romance`,
-  SCI_FI: `Sci-Fi`,
-  THRILLERS: `Thrillers`,
-};
-
-export {extend, Genres};
+export {extend};
